Reporter refactoring (#25)
* facade implementation of the reporter

* refactoring
